+++ base/tests/Common/ProcessImpl/TestStdPipeThread.cpp
@@ -25,7 +25,7 @@ namespace
         Common::ProcessImpl::StdPipeThread t(pipe.readFd());
         t.start();
         std::string expected_output("EXPECTED OUTPUT");
-        int ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
+        ssize_t ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
         ASSERT_EQ(ret, expected_output.size());
         pipe.closeWrite();
         // Don't request stop - wait for EOF to terminate the thread instead.
@@ -41,7 +41,7 @@ namespace
         t.setOutputLimit(limit);
         t.start();
         std::string expected_output("EXPECTED OUTPUT");
-        int ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
+        ssize_t ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
         ASSERT_EQ(ret, expected_output.size());
         pipe.closeWrite();
         // Don't request stop - wait for EOF to terminate the thread instead.
@@ -58,7 +58,7 @@ namespace
         t.setOutputLimit(limit);
         t.start();
         std::string expected_output("EXPECTED OUTPUT");
-        int ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
+        ssize_t ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
         ASSERT_EQ(ret, expected_output.size());
         pipe.closeWrite();
         // Don't request stop - wait for EOF to terminate the thread instead.
@@ -75,7 +75,7 @@ namespace
         t.setOutputLimit(limit);
         t.start();
         std::string expected_output("EXPECTED OUTPUT");
-        int ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
+        ssize_t ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
         ASSERT_EQ(ret, expected_output.size());
         pipe.closeWrite();
         // Don't request stop - wait for EOF to terminate the thread instead.
@@ -90,7 +90,7 @@ namespace
         const size_t limit = 50;
         t.setOutputLimit(limit);
         std::string expected_output("EXPECTED OUTPUT");
-        int ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
+        ssize_t ret = ::write(pipe.writeFd(), expected_output.c_str(), expected_output.size());
         ASSERT_EQ(ret, expected_output.size());
         pipe.closeWrite();
         // Note request for stop occurs before we request the thread to start
